br_investing/scrap/requirements.txt
google-cloud-bigquery-storage
beautifulsoup4